requirements.txt
pyodbc>=4.0.39
PyYAML>=6.0.1
flask>=2.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
flake8>=6.0.0
black>=23.0.0
mypy>=1.5.0